ama/pyproject.toml
[project]
name = "ama"
version = "0.1.0"
description = "Observability-first RDF vs LPG experiment scaffold for FinDER and FIBO."
requires-python = ">=3.12"
dependencies = [
  "datasets>=2.20.0",
  "python-dotenv>=1.0.1",
  "openai>=1.30.0",
  "pydantic>=2.7.0",
  "opik>=1.8.0",
  "neo4j>=5.20.0",
  "openai-agents>=0.1.0",
  "rdflib>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["framework", "experiments", "tracing", "debate"]

[tool.setuptools]
packages = ["framework", "experiments", "tracing", "debate"]
